Refactor uncheckedSource-related bazel queries
