+++ metals/src/main/scala/scala/meta/internal/metals/mbt/importer/BazelMbtImporter.scala
@@ -71,8 +71,7 @@ abstract class BazelMbtImporter(
         .run(queryEnv)
       targetsXmlDump = new BazelTargetsXmlDump(targetsXmlQueryOutput)
       srcs = targetsXmlDump.getLabels("srcs")
-      (genSrcOutputsByTarget, genSrcLabels) <- BazelQuery
-        .queryGenSrcOutputsByTarget(srcs, queryEnv)
+      (genSrcOutputsByTarget, genSrcLabels) <- queryGenSrcOutputsByTarget(srcs)
       filteredSrcs = srcs.map { case (t, labels) =>
         t -> labels.filterNot(genSrcLabels)
       }
@@ -265,6 +264,108 @@ abstract class BazelMbtImporter(
   private def extractScalaVersionFromLabel(label: String): Option[String] = {
     val versionPattern = """scala[_-]library[_-](\d+\.\d+\.\d+)""".r
     versionPattern.findFirstMatchIn(label).map(_.group(1))
+  }
+
+  /**
+   * For any entry in `srcsByTarget` that is a Bazel rule (not a real source
+   * file), runs `bazel cquery --output=files` on those rules and maps the
+   * output file paths to `uncheckedSources` for the owning target.
+   * Also returns the set of gen src labels so callers can exclude them from
+   * `sources`.
+   */
+  private def queryGenSrcOutputsByTarget(
+      srcsByTarget: Map[String, List[String]]
+  ): Future[(Map[String, List[String]], Set[String])] = {
+    val allSrcLabels =
+      srcsByTarget.values.flatten.filter(_.startsWith("//")).toSet
+    if (allSrcLabels.isEmpty) {
+      Future.successful((Map.empty[String, List[String]], Set.empty[String]))
+    } else {
+      queryRuleLabels(allSrcLabels)
+        .flatMap { ruleLabels =>
+          if (ruleLabels.isEmpty)
+            Future.successful(
+              (Map.empty[String, List[String]], Set.empty[String])
+            )
+          else
+            runBazelCqueryOutputsByLabel(ruleLabels).map { outputsByGenLabel =>
+              (
+                groupGenOutputsByTarget(
+                  srcsByTarget,
+                  ruleLabels,
+                  outputsByGenLabel,
+                ),
+                ruleLabels,
+              )
+            }
+        }
+        .recover { case e =>
+          scribe.warn(s"bazel-mbt: failed to query generated source outputs", e)
+          (Map.empty[String, List[String]], Set.empty[String])
+        }
+    }
+  }
+
+  private def queryRuleLabels(srcLabels: Set[String]): Future[Set[String]] =
+    BazelQuery(s"set(${srcLabels.mkString(" ")})", BazelQuery.OutputMode.Xml)
+      .run(queryEnv)
+      .map { xml =>
+        val srcFiles = new BazelTargetsXmlDump(xml).sourceFileLabels
+        srcLabels.filterNot(srcFiles)
+      }
+
+  private def groupGenOutputsByTarget(
+      srcsByTarget: Map[String, List[String]],
+      ruleLabels: Set[String],
+      outputsByGenLabel: Map[String, List[String]],
+  ): Map[String, List[String]] =
+    srcsByTarget.flatMap { case (target, srcs) =>
+      val genPaths = srcs
+        .filter(ruleLabels)
+        .flatMap(outputsByGenLabel.getOrElse(_, Nil))
+      Option.when(genPaths.nonEmpty)(target -> genPaths)
+    }
+
+  /** bazel cquery that returns output file paths grouped by label. */
+  private def runBazelCqueryOutputsByLabel(
+      labels: Set[String]
+  ): Future[Map[String, List[String]]] = {
+    val starlarkExpr =
+      """str(target.label) + "\t" + "\t".join([f.path for f in target.files.to_list()])"""
+    BazelQuery(
+      s"set(${labels.mkString(" ")})",
+      BazelQuery.OutputMode.Starlark,
+      extraArgs = List(s"--starlark:expr=$starlarkExpr"),
+      queryType = BazelQuery.QueryType.CQuery,
+    ).run(queryEnv)
+      .map { output =>
+        output.linesIterator
+          .map(_.trim)
+          .filter(_.nonEmpty)
+          .flatMap { line =>
+            line.split("\t") match {
+              case Array(rawLabel, paths @ _*) =>
+                val label =
+                  if (rawLabel.startsWith("@@//")) rawLabel.substring(2)
+                  else rawLabel
+                Some(label -> paths.flatMap(normalizeBazelOutputPath).toList)
+              case _ =>
+                scribe.warn(
+                  s"bazel-mbt: unexpected cquery starlark line: $line"
+                )
+                None
+            }
+          }
+          .toMap
+      }
+  }
+
+  /** Converts `bazel-out/<config>/bin/<rest>` to `bazel-bin/<rest>`. */
+  private def normalizeBazelOutputPath(path: String): Option[String] = {
+    val binMarker = "/bin/"
+    val idx = path.indexOf(binMarker)
+    if (idx < 0) None
+    else Some("bazel-bin/" + path.substring(idx + binMarker.length))
   }
 
   private def queryOutputBase(): Future[Option[Path]] = {
