+++ metals/src/main/scala/scala/meta/internal/metals/mbt/importer/BazelQuery.scala
@@ -45,7 +45,7 @@ object BazelQuery {
     case object CQuery extends QueryType("cquery")
   }
 
-  import OutputMode._, QueryType._
+  import OutputMode._
 
   def buildRuleKindsQuery(patterns: List[String]): BazelQuery = {
     val ps =
@@ -107,108 +107,6 @@ object BazelQuery {
       "scala_test", "java_test",
     )
 
-  /**
-   * For any entry in `srcsByTarget` that is a Bazel rule (not a real source
-   * file), runs `bazel cquery --output=files` on those rules and maps the
-   * output file paths to `uncheckedSources` for the owning target.
-   * Also returns the set of gen src labels so callers can exclude them from
-   * `sources`.
-   */
-  def queryGenSrcOutputsByTarget(
-      srcsByTarget: Map[String, List[String]],
-      queryEnv: Env,
-  )(implicit
-      ec: ExecutionContext
-  ): Future[(Map[String, List[String]], Set[String])] = {
-    val allSrcLabels =
-      srcsByTarget.values.flatten.filter(_.startsWith("//")).toSet
-    if (allSrcLabels.isEmpty) {
-      Future.successful((Map.empty[String, List[String]], Set.empty[String]))
-    } else {
-      // check if the labels correspond to source files, or rules
-      BazelQuery(s"set(${allSrcLabels.mkString(" ")})", Xml)
-        .run(queryEnv)
-        .flatMap { xml =>
-          val targetsXmlDump = new BazelTargetsXmlDump(xml)
-          val srcFiles = targetsXmlDump.sourceFileLabels
-          val ruleLabels = allSrcLabels.filterNot(srcFiles)
-          if (ruleLabels.isEmpty) {
-            Future.successful(
-              (Map.empty[String, List[String]], Set.empty[String])
-            )
-          } else {
-            runBazelCqueryOutputsByLabel(ruleLabels, queryEnv).map {
-              outputsByGenLabel =>
-                val genSrcOutputsByTarget = srcsByTarget.flatMap {
-                  case (target, srcs) =>
-                    val genPaths = srcs
-                      .filter(ruleLabels)
-                      .flatMap(outputsByGenLabel.getOrElse(_, Nil))
-                    Option.when(genPaths.nonEmpty)(target -> genPaths)
-                }
-                (genSrcOutputsByTarget, ruleLabels)
-            }
-          }
-        }
-        .recover { case e =>
-          scribe.warn(s"bazel-mbt: failed to query generated source outputs", e)
-          (Map.empty[String, List[String]], Set.empty[String])
-        }
-    }
-  }
-
-  /**
-   * Single `bazel cquery --output=starlark` call that returns output file paths
-   * grouped by label.
-   */
-  def runBazelCqueryOutputsByLabel(
-      labels: Set[String],
-      queryEnv: Env,
-  )(implicit ec: ExecutionContext): Future[Map[String, List[String]]] = {
-    // Emit one tab-separated line per target: "//pkg:name\tpath1\tpath2..."
-    val starlarkExpr =
-      """str(target.label) + "\t" + "\t".join([f.path for f in target.files.to_list()])"""
-    BazelQuery(
-      s"set(${labels.mkString(" ")})",
-      Starlark,
-      extraArgs = List(s"--starlark:expr=$starlarkExpr"),
-      queryType = CQuery,
-    ).run(queryEnv)
-      .map { output =>
-        val result = output.linesIterator
-          .map(_.trim)
-          .filter(_.nonEmpty)
-          .flatMap { line =>
-            line.split("\t") match {
-              case Array(rawLabel, paths @ _*) =>
-                val label =
-                  if (rawLabel.startsWith("@@//")) rawLabel.substring(2)
-                  else rawLabel
-                val normalized = paths.flatMap(normalizeBazelOutputPath).toList
-                Some(label -> normalized)
-              case _ =>
-                scribe.warn(
-                  s"bazel-mbt: unexpected cquery starlark line: $line"
-                )
-                None
-            }
-          }
-          .toMap
-        result
-      }
-      .recover { case e =>
-        scribe.warn(s"bazel-mbt: cquery starlark query failed", e)
-        Map.empty
-      }
-  }
-
-  /** Converts `bazel-out/<config>/bin/<rest>` to `bazel-bin/<rest>`. */
-  private def normalizeBazelOutputPath(path: String): Option[String] = {
-    val binMarker = "/bin/"
-    val idx = path.indexOf(binMarker)
-    if (idx < 0) None
-    else Some("bazel-bin/" + path.substring(idx + binMarker.length))
-  }
 }
 
 case class BazelQuery(
