+++ tests/unit/src/test/scala/tests/mbt/MbtBuildServerLspSuite.scala
@@ -615,7 +615,6 @@ class MbtBuildServerLspSuite
     } yield ()
   }
 
-
   test("mbt-uncheckedSources-compiles-with-gitignore-sources") {
     cleanWorkspace()
     val mbtJson =
